perf(cli): slim entry so secrets-broker reads skip the commander graph (RUSH-2335)
Split the CLI entry so dist/index.js statically imports only the
lib/secrets/sync-commands leaf and dispatches the argv fast paths
(__secrets-*, __vault-age-helper, __shim, __daemon-run, __daemon-tick)
before loading the full CLI via await import('./bootstrap.js').

Cold node dist/index.js __secrets-ping drops from ~160-180ms to ~40-60ms
on this host (bare-node floor ~10ms). Position guards and the sync-commands
leaf invariant are tightened in agent.test.ts.
