+++ apps/cli/.changelog/next/RUSH-2335-entry-split.md
@@ -1,0 +1,1 @@
+- **Synchronous secrets-broker reads no longer pay ~140ms of CLI bootstrap (RUSH-2335).** `dist/index.js` is now a slim shell that statically imports only the leaf `lib/secrets/sync-commands.js` and dispatches `__secrets-get` / `__secrets-ping` / `__secrets-lock` (plus `__vault-age-helper`, `__shim`, `__daemon-run`, `__daemon-tick`) before loading the full commander graph. Everything else arrives via `await import('./bootstrap.js')`. Cold `node dist/index.js __secrets-ping` was measured at ~160–180ms against a ~22ms bare-node floor; the leaf-only path evaluates in roughly one bare-node spawn plus the agent handler. Source: `apps/cli/src/index.ts`, `apps/cli/src/bootstrap.ts`.
